fix for PR comments
